class to precompute fft structures

+++ fhe/BigReal.cpp
@@ -154,7 +154,7 @@ void to_BigTorus(BigTorusRef dest, const BigReal &a, UINT64 lshift_bits, UINT64 
     //...
     int64_t asize = abs(a.value->_mp_size);
     for (int i = 0; i <= out_precision_limbs; i++) {
-        if (a.nblimbs - i - 1 < asize && a.nblimbs - i - 1 >= 0)
+        if (int64_t(a.nblimbs - i - 1) < asize && int64_t(a.nblimbs - i - 1) >= 0)
             tmp[out_precision_limbs - i] = a.value->_mp_d[a.nblimbs - 1];
         else
             tmp[out_precision_limbs - i] = 0;
